[Disco] Improve RCCL Integration Logic
This PR makes the following changes:
- Move `runtime/disco/ccl` to `runtime/disco/nccl` as RCCL itself uses
  NCCL as a prefix to its APIs;
- Consolidate `utils.h` to `nccl.cc` to avoid potential misuse of the
  private header file;
- Reduce the number of macros used in `nccl.cc`.
